### 2022-02-02 #4127 @whedon
[PRE REVIEW]: tesliper: a theoretical spectroscopist's little helper
**Submitting author:** @mishioo (<a href="http://orcid.org/0000-0001-7884-8982">Michał M. Więcław</a>)
**Repository:** <a href="https://github.com/mishioo/tesliper" target ="_blank">https://github.com/mishioo/tesliper</a>
**Version:** 0.9.0
