### 2022-02-14 #4127 @rkurchin
Sounds great, thanks for the suggestion!

@arepstein, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
