### 2022-02-14 #4164 @whedon
[REVIEW]: tesliper: a theoretical spectroscopist's little helper
**Submitting author:** @mishioo (<a href="http://orcid.org/0000-0001-7884-8982">Michał M. Więcław</a>)
**Repository:** https://github.com/mishioo/tesliper
**Branch with paper.md** (empty if default branch):
**Version:** 0.9.3
