### 2022-02-18 3bf1dac
Added a test for `Conformers.arrayed("filenames")`.
