### 2022-02-22 5ee94a5
Added "strict" kwarg to the `Conformers.arrayed()` method.
This new keyword argument controls how unexpected additional keyword
arguments are treated.
