### 2022-02-22 04e21d1
Updated `Conformers.arrayed()` docstring.
