### 2022-02-22 d912abd
Added `Tesliper.temperature` property.
It holds system's temperature value. It is passed to data array classes
constructed via subscription mechanism.
