### 2022-02-22 53ce0fe
Added optional "temperature" parameter to `Tesliper.__init__()`.
This change is introduced for easier integration with serialization
mechanism. It mey be reversed in the next major release.
