### 2022-02-22 f35a7cf
Updated serialization to include `Tesliper.temperature`.
