+++ tesliper/writing/serializer.py
@@ -130,7 +130,11 @@ class ArchiveWriter:
     def write(self, obj: "tesliper.Tesliper"):
         with self:
             self._write_arguments(
-                obj.input_dir, obj.output_dir, obj.wanted_files, obj.quantum_software
+                obj.input_dir,
+                obj.output_dir,
+                obj.wanted_files,
+                obj.temperature,
+                obj.quantum_software,
             )
             self._write_parameters(obj.parameters)
             self._write_conformers(obj.conformers)
@@ -146,6 +150,7 @@ class ArchiveWriter:
         input_dir: Union[Path, str] = None,
         output_dir: Union[Path, str] = None,
         wanted_files: Iterable[str] = None,
+        temperature: float = None,
         quantum_software: str = None,
     ):
         with self.root.open("arguments.json", mode="w") as handle:
@@ -155,6 +160,7 @@ class ArchiveWriter:
                         "input_dir": str(input_dir) if input_dir else None,
                         "output_dir": str(output_dir) if output_dir else None,
                         "wanted_files": list(wanted_files) if wanted_files else None,
+                        "temperature": temperature,
                         "quantum_software": quantum_software or None,
                     }
                 )
