### 2022-02-22 86ab3a3
Updated `Tesliper.get_average_spectrum()`.
Now it allows to override `Tesliper.temperature` value with the optional
*temperature* parameter.
