### 2022-02-22 fea78f8
Added a fallback for spectra in `Tesliper.get_averaged_spectrum()`.
