### 2022-02-22 15a713c
Added basic tests for `Tesliper.get_averaged_spectrum()`.
