+++ test/integration/test_tesliper.py
@@ -223,3 +223,22 @@ def test_subscription_temperature(value):
     tslr.temperature = value
     array = tslr["zpe"]
     assert array.t == value
+
+
+def test_get_averaged_spectrum(calculated):
+    spc = calculated.get_averaged_spectrum("ir", "zpe")
+    assert spc.genre == "ir"
+    assert spc.averaged_by == "zpe"
+
+
+def test_get_averaged_calculate(calculated, extracted):
+    e = extracted.get_averaged_spectrum("ir", "zpe")
+    c = calculated.get_averaged_spectrum("ir", "zpe")
+    assert e.values.tolist() == c.values.tolist()
+
+
+def test_get_average_raises_missing(tesliper):
+    assert "ir" not in tesliper.spectra
+    assert not tesliper.conformers.has_genre("dip")
+    with pytest.raises(ValueError):
+        tesliper.get_averaged_spectrum("ir", "zpe")
