### 2022-02-22 4030d98
Corrected `len()` behavior with `Spectra` instances.
