### 2022-02-23 82d6ca5
Added "Temperature" entry in GUI.
The entry is connected with `Tesliper.temperature` attribute.
Its change is reflected in GUI's views in real-time.
