### 2022-02-23 7792d07
Bound `live_preview_callback()` to <<KeptChanged>> event.
Now average spectrum is updated whenever kept conformers change.
