### 2022-02-23 6b6b682
Restricted temperature to values above abs zero.
Change is applied to `Tesliper.temperature` and GUI only.
