### 2022-02-23 3dc9007
Added section about changing temperature in GUI docs.
