### 2022-02-24 #4164 @alejandrogallo
Hi @rkurchin, is it ok for reviewers to issue pull-requests for things like typos or should we open issues for these?
