### 2022-02-27 bc55c23
Moved requirements to `setup.py` file.
