### 2022-02-27 c569217
Updated `appveyor.yml` to use `setup.py` requirements.
