### 2022-02-27 9464c2d
Changed location of metadata variables (version, author).
This allows to install from the local `setup.py`.
Previously an import of these variables was causing an ImportError.
