### 2022-02-27 8e80b41
Closes #4; Added `tesliper-gui` entry point.
