### 2022-02-27 e91af2f
Closes #2; Configured matrix CI testing.
