### 2022-02-27 26c15fe
Corrected import of dict views.
`ItemsView`, `KeysView`, `ValuesView` should be imported from
`collections.abc`, where they are defined. Python 3.10 changes how they
are imported to `collections`, causing an exception on `from collections
import ItemsView, ...`.
