+++ .appveyor.yml
@@ -1,8 +1,11 @@
 version: 0.9.0-{branch}-{build}
 
+stack: python 3
 
 environment:
   matrix:
+    - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu
+    - APPVEYOR_BUILD_WORKER_IMAGE: macOS
     - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       PYTHON: "C:\\Python36"
     - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
@@ -13,9 +16,6 @@ environment:
       PYTHON: "C:\\Python39"
     - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       PYTHON: "C:\\Python310"
-    - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu
-      stack: python 3
-    - APPVEYOR_BUILD_WORKER_IMAGE: macOS
   COVERALLS_REPO_TOKEN:
     secure: 64bL5OipIuqAwg0nGwbra4tIVBheT7jwIKgDuGuV2SefqhwLATu0dSgc/w+ecUE0
 
