### 2022-02-27 1a141ff
Separated note about `tkinter` dependency.
