### 2022-02-27 4b3242c
Moved `.temperature` to `Conformers`.
It's easier to fully integrate this new attribute with
`Conformers.arrayed()` when it is defined in the `Conformers` class.
`Tesliper` class still provides `.temperature`, but only as a proxy.

Additionally:
- corrected room temperature to 298.15K (25 C) where it was wrong;
- corrected `Conformers` docstrings;
- corrected `.temperature` tests.
