### 2022-02-27 ce80166
Added section about temperature in scripting tutorial.
