### 2022-02-27 a13b78c
Supplemented room temperature information.
