+++ docs/source/gui.rst
@@ -160,9 +160,9 @@ The ``Energies and structure`` section also allows you to specify the temperatur
 of the studied system. This parameter is important for calculation of the Boltzmann
 distribution of conformers, which is used to estimate conformers' population
 and average conformers' spectra. The default value is the room temperature,
-expressed as :math:`298.15 \mathrm{Kelvin}`. Changing this value will trigger automatic
-recalculation of ``Min. Boltzmann factor`` and ``Population`` values, and average
-spectra will be redrawn.
+expressed as :math:`298.15\ \mathrm{Kelvin}` (:math:`25.0^{\circ}\mathrm{C}`).
+Changing this value will trigger automatic recalculation of ``Min. Boltzmann factor``
+and ``Population`` values, and average spectra will be redrawn.
 
 .. image:: _static/screenshots/16426977988.png
 
