### 2022-02-27 b9867bb
Supplemented "Participating in code" section in READEME.md
