### 2022-02-27 9821455
Updated READEME.md and docs with new GUI invocation instructions.
