+++ tesliper/_metadata.py
@@ -1,5 +1,5 @@
 __author__ = "Michał M. Więcław"
-__version_info__ = (0, 9, 0)
+__version_info__ = (0, 9, 1)
 __version__ = ".".join(map(str, __version_info__))
 # approach coined from suggestions in this thread:
 # https://stackoverflow.com/q/458550/11416569
