+++ .appveyor.yml
@@ -1,4 +1,4 @@
-version: 0.9.0-{branch}-{build}
+version: 0.9.1-{branch}-{build}
 
 stack: python 3
 
