### 2022-02-28 6c31857
Updated contributing guidelines with typing info.
Also reworded "Participating in code" section a little.
