### 2022-02-28 7a9e517
Updated `setup.py` file.
Added Python 3.10 as supported.
Added required Python version.
Changed pyinstaller command to use descriptive flags.
