+++ setup.py
@@ -25,17 +25,17 @@ class BuildBinaryCommand(distutils.cmd.Command):
     def run(self):
         command = [
             "pyinstaller",
-            "-y",
-            "-F",
-            "-i",
+            "--noconfirm",
+            "--onefile",
+            "--icon",
             ".\\tesliper\\tesliper.ico",
-            "-n",
+            "--name",
             "Tesliper",
-            "-p",
+            "--paths",
             ".",
             "--add-data",
             ".\\tesliper\\tesliper.ico;tesliper",
-            "-w",
+            "--noconsole",
             ".\\bin\\tesliper_gui.py",
         ]
         subprocess.run(command)
@@ -71,6 +71,7 @@ setup(
     packages=find_packages(),
     install_requires=["numpy", "openpyxl"],
     extras_require=extras_require,
+    python_requires=">=3.6",
     scripts=["bin/tesliper_gui.py"],
     entry_points={"console_scripts": ["tesliper-gui=tesliper.gui:run"]},
     package_data={"tesliper": ["tesliper.ico"]},
@@ -81,6 +82,7 @@ setup(
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Topic :: Scientific/Engineering :: Chemistry",
         "Development Status :: 4 - Beta",
     ],
