+++ tesliper/_metadata.py
@@ -3,3 +3,5 @@ __version_info__ = (0, 9, 1)
 __version__ = ".".join(map(str, __version_info__))
 # approach coined from suggestions in this thread:
 # https://stackoverflow.com/q/458550/11416569
+
+__all__ = ["__author__", "__version_info__", "__version__"]
