### 2022-03-04 #4164 @arepstein
@rkurchin I don't think the paper compiled correctly. Is that an issue with the repo that I should raise or is it with JOSS? Thanks
