### 2022-03-04 #4164 @mishioo
The paper is on a separate branch "joss", but maybe I'll just merge it to "master" to make its compilation straightforward.
