### 2022-03-14 #4164 @arepstein
Hi @rkurchin and @mishioo, I've completed my checklist and I think Tesliper meets all of the criteria for publication in JOSS. The scientific paper establishes a clear need for a free, open-source spectroscopy tool that includes conformers and it explains why Tesliper addresses that need. I tested basic functionality with some of my own Gaussian output files both using the GUI and a Python notebook and I found the documentation to be  clear and easy to follow.

I have a few small comments. I can put them in issues if that would be more helpful, but they're not quite issues so I thought I'd write them here first.

- Some errors in Gaussian output files cause errors in glassware but not the parser. It might be nice to include warnings for bad log files or at least a log file lacking "Normal termination."
- I think the recursive parsing functionality is important, as running each conformer in its own directory is common, but I don't think that's available in the GUI.
- I would consider adding test coverage guidelines for contributors, such as with [coveralls.io](https://coveralls.io/).
- In my work, I've had some issues with RMSD calculations for different algorithms and have often needed to try a few to see which work best for a given molecule. I would consider adding some additional options in addition to Kabsch. [Pymatgen's Molecule Matcher](https://github.com/materialsproject/pymatgen/blob/c37ba66cce99e93557b5e44d697059e2c93c0869/pymatgen/analysis/molecule_matcher.py) has a few that I would consider adding. The InchiMolAtomMapper is particularly nice when it works because I believe the atomic ordering does not need to be maintained between the conformers.

Thank you for writing this, it seems like it will be quite helpful to a wide range of research!
