### 2022-03-14 mishioo/tesliper#5 @arepstein
GUI installation error with ZSH
Hi, I had a small issue installing the gui with a zsh shell. I needed to escape the brackets, i.e. `pip install tesliper\[gui\]` instead of `pip install tesliper[gui]`.
