### 2022-03-15 mishioo/tesliper#5 @mishioo
I'm not familiar with zsh, but [this stackexchange answer](https://apple.stackexchange.com/a/427617) explains that this is due to how zsh resolves `some[thing]` syntax, so the issue is not related to `tesliper`. Anyway, I will add a note about this for zsh users in the "Installation" section of docs.
