### 2022-03-15 mishioo/tesliper#5 @arepstein
Thanks! Yes I see that's not related tesliper, but with zsh as the default for Mac users I think a note makes sense.
