### 2022-03-15 #4164 @mishioo
Thank you for your insight, @arepstein! And also for words of appreciation, as I was a bit concerned if this project will be useful enough for a broader public, but I really hope so! Please, let me address you comments below:

- About errors: if you mean just adding a `logging.info` or a `logging.warning` at the end of parsing an output file that has an error or that ended abruptly, then it's easily doable. There is already a `logging.info` when parsing unexpectedly stops, but maybe changing it to `warning` would be better. Anyway, I'm not quite sure about errors in glassware part; when does they happen? If you feel this doesn't suppose to happen, then I would appreciate if you'd open an issue with more details, and, ideally, with an example log file.
- About recursive parsing: I'm glad you mentioned this. the group that have done initial testing didn't need this feature, but I was not sure about other groups workflow. I'll add it to GUI.
- About coverage: valid point, I'll update relevant READEME section.
- About RMDS: that's a tougher one... I am aware of the limitations of Kabsch/RMSD approach, especially when rotation of symmetrical groups like -Me or -Ph comes into play. I was thinking about expanding the conformers comparison functionality with other methods myself, but on the other hand, I feel like adding a Pymatgen, Open Babel, and SciPy as dependencies would be quite an overkill for such a humble project as `tesliper`... I think it's a valid route of development, but I'd prefer to leave it for now and come back to it at a later time, after some more research on my part. Is this acceptable?
