### 2022-03-15 #4164 @arepstein
- For the errors, I do think a `warning` could be more helpful for a user debugging. But I would also leave that up to you, I don't think it's an absolutely necessary change. The example I used was deleting several frequencies as if the job timed out and that's when I got an error in the glassware -- it is a very expected error because the output file was wrong, so I think the error was supposed to happen and don't think an issue is necessary.
- For the RMSDs, I completely understand about not wanting to add large dependencies and agree that it makes sense as a future addition and is not necessary for publication.
