### 2022-03-23 a1c52a4
Closes #5. Added a note about installation in zsh.
