### 2022-03-23 36159b7
Supplemented logging on parsing abnormally terminated files.
Now warning will be issued when `Soxhlet` finds any abnormally
terminated files, but only once per `Soxhlet.extract()` call.
Unexpected `StopIteration` encountered in `GaussianParser` is now
signalized with debug message rather than an info.
