### 2022-03-23 0ac6c81
Removed ticks from section title so it renders more nicely.
