### 2022-03-23 #4164 @mishioo
So happy to see the first green light for `tesliper`'s publication! :)
Also, an update from me, as I was finally able to introduce changes we talked about. @arepstein could you please confirm if they satisfy your suggestions?

- I've added a note about [installation in zsh](https://tesliper.readthedocs.io/en/latest/installation.html#install-from-terminal), as we discussed in [this issue](https://github.com/mishioo/tesliper/issues/5).
- I've added the recursive extraction to GUI & updated [the relevant part of docs](https://tesliper.readthedocs.io/en/latest/gui.html).
- I've updated testing guidelines in [relevant part of READEME](https://github.com/mishioo/tesliper#participating-in-code) with coverage info.
- I've resolved the matter of abnormal termination warnings by moving it up by one level (from `GaussianParser` to `Soxhlet` class) -- now a warning will be issued if any abnormally terminated files are detected in a single batch of extraction (i.e. only once for `extract()` call). I think this is a sane middle ground. Relevant changes [in this commit](https://github.com/mishioo/tesliper/commit/36159b7e7329b293c19b342226f7df7cb476b1ee).

I didn't issue a new release yet, I decided to wait with this for possible further comments from @alejandrogallo.
