### 2022-04-04 62aefec
Corrected software comparison.
Changed CDspecTech's mistakenly checked characteristics.
