### 2022-04-10 #4164 @alejandrogallo
I have concluded my review of the most recent version. I have to thank also @arepstein for raising quite relevant points like the recursive parsing.

I am satisfied with the current status of the package (da2b096db6adf66e64e6fe0923f40f434d64c976) and **I recommend this package for publication at JOSS**.

The author @mishioo has done a great job addressing the points raised and further improving the documentation, the CI infrastructure, and thus the overall quality. I have to praise also the clear writing of the introductions with pertinent references to related software that I could therefore compare it to. I think `tesliper` fills a very relevant gap in the spectroscopy ecosystem and can now serve as a robust starting point for other projects.

Furthermore, the current state of the documentation for typical workflows, GUI program, scripting and advanced usage ensures that even if shortcomings appear in the future, the maintenance of the software is realistic and thus it represents a relevant long-term addition to the community.

The documentation is also clearly written for material scientists or quantum chemists in general as myself that are somewhat external to the field and need guidelines to get started quickly. Thus, the presented examples are well written, can be followed and work as intended in my testing.

Upon further reviewing of the code, I think that extensions to [`pymatgen`](https://pymatgen.org/), [`babel`](https://www.openbabel.org/), [`ase`](http://wiki.fysik.dtu.dk/ase) or related software packages will probably happen in the future and they are realistically implemented, but I concur with @arepstein that these are not necessary extensions for the current publication and can be left for future enhancements. Indeed, the clear structuring of the core package enables it.

I thank the author @mishioo for his outstanding work and the editor @rkurchin for her free software and open science commitment and excellent editorial task.
