### 2022-04-11 #4164 @mishioo
Thank you @alejandrogallo and @arepstein that you have found time for this review. I'm very happy to see `tesliper` recommended for publication!

The commit mentioned by @alejandrogallo is the latest release, issued in the meantime with version number 0.9.3.
I have created a Zenodo archive for it, the assigned DOI is [10.5281/zenodo.6448370](https://doi.org/10.5281/zenodo.6448370).
Is there anything else I should do now, @rkurchin?
