### 2022-04-11 #4164 @rkurchin
A few minor editorial suggestions:
* line 34: singular/plural disagreement – perhaps rephrase to "...adoption of one or more of the software packages mentioned above..." or "...adoption of one or more piece of the aforementioned software..."
* line 35: parallelism - suggest change to "...incomplete, or unable to be done in an automated fasion."
* 37: hyphenate "time-consuming"
