### 2022-04-12 #4164 @mishioo
@rkurchin, @alejandrogallo, @arepstein, @arfon, thank you all for conducting this review, I'm extremely happy to see my software published in JOSS! :)
